+++ src/infrastructure/providers/index.ts
@@ -1,2 +1,3 @@
+export { AwsGeocoder } from './AwsGeocoder';
 export { BrowserGeolocationProvider } from './BrowserGeolocationProvider';
 export { MockGeolocationProvider } from './MockGeolocationProvider';